+++ lib/ggity/stat.ex
@@ -1,0 +1,41 @@
+defmodule GGity.Stat do
+  @moduledoc false
+
+  @type dataset :: list(map())
+
+  @doc false
+  @spec identity(dataset(), map()) :: {dataset(), map()}
+  def identity(data, mapping), do: {data, mapping}
+
+  # TODO: generalize this function so that it know which aesthetics to interact
+  # based on the provided mapping and whether or not the associated scale is discrete
+  @doc false
+  @spec count(dataset(), map()) :: {dataset(), map()}
+  def count(data, mapping) do
+    permutations =
+      for row <- data,
+          uniq: true,
+          do: {row[mapping[:x]], row[mapping[:fill]], row[mapping[:alpha]]}
+
+    stat =
+      Enum.reduce(permutations, [], fn {x_value, fill_value, alpha_value}, stat ->
+        [
+          Map.new([
+            {mapping[:x], x_value},
+            {mapping[:fill], fill_value},
+            {mapping[:alpha], alpha_value},
+            {:count,
+             Enum.count(data, fn row ->
+               {row[mapping[:x]], row[mapping[:fill]], row[mapping[:alpha]]} ==
+                 {x_value, fill_value, alpha_value}
+             end)}
+          ])
+          | stat
+        ]
+      end)
+      |> Enum.sort_by(fn row -> row[mapping[:x]] end)
+
+    mapping = Map.put(mapping, :y, :count)
+    {stat, mapping}
+  end
+end